Refactor the compiler as a library
