+++ Cargo.toml
@@ -1,8 +1,15 @@
 [package]
-name = "zephyr"
+name = "zephyr-lang"
 version = "0.1.0"
 authors = ["CharlyCst <charly.castes@epfl.ch>"]
 edition = "2018"
+
+[lib]
+name = "zephyr"
+
+[[bin]]
+name = "zephyr"
+path = "src/bin/main.rs"
 
 [dependencies]
