+++ src/ast/ast.rs
@@ -1,7 +1,8 @@
-pub use crate::driver::ModulePath;
+use std::fmt;
+
+pub use crate::ctx::ModulePath;
 use crate::error::Location;
 use crate::mir::Value as MirValue;
-use std::fmt;
 
 ////// Zephyr AST nodes //////
