--- src/driver/mod.rs
@@ -1,10 +1,0 @@
-mod utils;
-mod ctx;
-mod resolver;
-mod known_functions;
-
-pub use utils::{PublicDeclarations, ModuleDeclarations, ModulePath};
-pub use ctx::Ctx;
-pub use known_functions::KnownFunctions;
-pub use resolver::{Resolver, StandardResolver};
-
