+++ src/lib.rs
@@ -1,0 +1,14 @@
+//! # The Zephyr compiler
+
+mod ast;
+mod ctx;
+mod error;
+mod hir;
+mod mir;
+mod resolver;
+mod wasm;
+
+pub use ctx::{Ctx, ModuleKind, ModulePath, PreparedFile};
+pub use error::ErrorHandler;
+pub use resolver::{Resolver, StandardResolver};
+pub use ast::Kind as FileKind;
