+++ src/resolver/mod.rs
@@ -1,0 +1,5 @@
+mod resolver;
+mod standard_resolver;
+
+pub use resolver::Resolver;
+pub use standard_resolver::StandardResolver;
